pyproject.toml
[project]
name = "slack-integration"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-generativeai>=0.8.5",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "requests>=2.32.4",
    "uvicorn>=0.35.0",
    "sqlalchemy>=2.0.0",
    "pandas>=2.3.1",
]
